# docs/swagger.yaml
definitions:
  aiapi.ChatCompletionRequest:
    properties:
      frequency_penalty:
        type: number
      max_tokens:
        type: integer
      messages:
        items:
          $ref: '#/definitions/openai.ChatCompletionMessage'
        type: array
      model:
        type: string
      role:
        description: 角色
        type: string
      stream:
        type: boolean
      temperature:
        type: number
    type: object
  aiapi.ImageRequest:
    properties:
      model:
        default: dall-e2
        description: dall-e2/stable-diffusion
        type: string
      "n":
        default: 1
        type: integer
      prompt:
        type: string
    type: object
...
  openai.ChatCompletionMessage:
    properties:
      content:
        type: string
      name:
        description: |-
          This property isn't in the official documentation, but it's in
          the documentation for the official library for python:
          - https://github.com/openai/openai-python/blob/main/chatml.md
          - https://github.com/openai/openai-cookbook/blob/main/examples/How_to_count_tokens_with_tiktoken.ipynb
        type: string
      role:
        type: string
    type: object
  payapi.PreCreateRequest:
    properties:
      payType:
        default: alipay
        description: 支付类型，目前只支持alipay
        type: string
      productType:
        description: 商品类型，1-10元,2-30元,3-100元
        type: integer
    required:
    - productType
    type: object
  payapi.PreCreateResponse:
    properties:
      orderId:
        description: 订单id
        type: string
      qrCode:
        description: 二维码串
        type: string
    type: object
  user.ChangePasswordRequest:
    properties:
      newPassword:
        type: string
      subjectName:
        type: string
      subjectType:
        default: 101
        description: 可选字段，默认为userapi.ChangePasswordMailCode
        type: integer
      verificationCode:
        type: string
    required:
    - newPassword
    - subjectName
    - verificationCode
    type: object
...
  user.SendPasswordVerificationCodeRequest:
    properties:
      subjectName:
        type: string
      subjectType:
        default: 101
        description: 可选字段，默认为userapi.ChangePasswordMailCode
        type: integer
    required:
    - subjectName
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/openai/v1/audio:
    post:
      parameters:
      - description: 音频文件
        in: formData
        name: audio
        required: true
        type: file
      - description: model
        in: query
        name: model
        type: string
      - description: language
        in: query
        name: language
        type: string
      - description: prompt
        in: query
        name: prompt
        type: string
      - description: temperature
        in: query
        name: temperature
        type: number
      responses: {}
      summary: 音频转文字
      tags:
      - OpenAI相关接口
  /api/v1/openai/v1/chat/completions:
    post:
      parameters:
      - description: openai请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/aiapi.ChatCompletionRequest'
      responses: {}
      summary: 对话
      tags:
      - OpenAI相关接口
  /api/v1/openai/v1/image:
    post:
      parameters:
      - description: openai请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/aiapi.ImageRequest'
      responses: {}
      summary: 生成图片
      tags:
      - OpenAI相关接口
...
  /api/v1/pay/pre_create:
    post:
      parameters:
      - description: 订单请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/payapi.PreCreateRequest'
      responses:
        "200":
          description: 创建订单返回参数
          schema:
            $ref: '#/definitions/payapi.PreCreateResponse'
      summary: 预创建交易订单
      tags:
      - 支付相关接口
...
  /api/v1/user/change_password:
    post:
      parameters:
      - description: 发送验证码请求参数(默认为email)
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/user.ChangePasswordRequest'
      responses: {}
      summary: 用户修改密码
      tags:
      - 用户相关接口
  /api/v1/user/change_password/verify/code:
    post:
      parameters:
      - description: 发送验证码请求参数(默认为email)
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/user.SendPasswordVerificationCodeRequest'
      responses: {}
      summary: 在忘记密码时发送验证码
      tags:
      - 用户相关接口
...
swagger: "2.0"